+++ clevertap-core/src/main/java/com/clevertap/android/sdk/inapp/data/InAppResponseAdapter.kt
@@ -37,9 +37,8 @@ class InAppResponseAdapter(
         }
     }
 
-    val preloadImage: List<String>
-
-    val allImages: List<String>
+    val preloadImages: List<String>
+    val preloadGifs: List<String>
 
     val legacyInApps: Pair<Boolean, JSONArray?> = responseJson.safeGetJSONArray(Constants.INAPP_JSON_RESPONSE_KEY)
 
@@ -50,21 +49,22 @@ class InAppResponseAdapter(
     val appLaunchServerSideInApps: Pair<Boolean, JSONArray?> = responseJson.safeGetJSONArray(Constants.INAPP_NOTIFS_APP_LAUNCHED_KEY)
 
     init {
-        val list = mutableListOf<String>()
+        val imageList = mutableListOf<String>()
+        val gifList = mutableListOf<String>()
 
-        fetchMediaUrls(legacyInApps, list)
-        fetchMediaUrls(clientSideInApps, list)
-        fetchMediaUrls(appLaunchServerSideInApps, list)
+        //fetchMediaUrls(legacyInApps, list)
+        //fetchMediaUrls(appLaunchServerSideInApps, list)
+        fetchMediaUrls(clientSideInApps, imageList, gifList)
 
-        preloadImage = list
-
-        val allImagesList = list.toMutableList()
-        fetchMediaUrls(serverSideInApps, allImagesList)
-
-        allImages = allImagesList
+        preloadImages = imageList
+        preloadGifs = imageList
     }
 
-    private fun fetchMediaUrls(data: Pair<Boolean, JSONArray?>, list: MutableList<String>) {
+    private fun fetchMediaUrls(
+        data: Pair<Boolean, JSONArray?>,
+        imageList: MutableList<String>,
+        gifList: MutableList<String>
+    ) {
         if (data.first) {
             data.second?.iterator<JSONObject> { jsonObject ->
                 val portrait = jsonObject.optJSONObject(Constants.KEY_MEDIA)
@@ -74,7 +74,11 @@ class InAppResponseAdapter(
                         .initWithJSON(portrait, Configuration.ORIENTATION_PORTRAIT)
 
                     if (portraitMedia != null && portraitMedia.mediaUrl != null) {
-                        list.add(portraitMedia.mediaUrl)
+                        if (portraitMedia.isImage) {
+                            imageList.add(portraitMedia.mediaUrl)
+                        } else if (portraitMedia.isGIF) {
+                            gifList.add(portraitMedia.mediaUrl)
+                        }
                     }
                 }
                 val landscape = jsonObject.optJSONObject(Constants.KEY_MEDIA_LANDSCAPE)
@@ -83,7 +87,11 @@ class InAppResponseAdapter(
                             .initWithJSON(landscape, Configuration.ORIENTATION_LANDSCAPE)
 
                     if (landscapeMedia != null && landscapeMedia.mediaUrl != null) {
-                        list.add(landscapeMedia.mediaUrl)
+                        if (landscapeMedia.isImage) {
+                            imageList.add(landscapeMedia.mediaUrl)
+                        } else if (landscapeMedia.isGIF) {
+                            gifList.add(landscapeMedia.mediaUrl)
+                        }
                     }
                 }
             }