+++ clevertap-core/src/test/java/com/clevertap/android/sdk/inapp/images/preload/InAppImagePreloaderExecutorsTest.kt
@@ -13,6 +13,9 @@ import kotlin.test.assertEquals
 class InAppImagePreloaderExecutorsTest {
 
     private val mockBitmap = mockk<Bitmap>()
+    private val byteArray = ByteArray(10) { pos ->
+        pos.toByte()
+    }
     private val inAppResourceProvider = mockk<InAppResourceProvider>()
     private val executors = MockCTExecutors()
 
@@ -43,6 +46,30 @@ class InAppImagePreloaderExecutorsTest {
             val url = urls[count]
             verify {
                 inAppResourceProvider.fetchInAppImage(url)
+            }
+        }
+        assertEquals(urls.size, successUrls.size)
+    }
+
+    @Test
+    fun `preload gifs fetches gif from all urls`() {
+        val urls = mutableListOf("a", "b", "c", "d", "e", "f", "g", "h", "i", "j", "k")
+        val successUrls = mutableListOf<String>()
+
+        for (url in urls) {
+            every {
+                inAppResourceProvider.fetchInAppGif(url)
+            } returns byteArray
+        }
+
+        inAppImagePreloader.preloadGifs(urls) { url ->
+            successUrls.add(url)
+        }
+
+        for (count in 0 until urls.size) {
+            val url = urls[count]
+            verify {
+                inAppResourceProvider.fetchInAppGif(url)
             }
         }
         assertEquals(urls.size, successUrls.size)